Update elastic version
WMF's APT retpo doesn't serve ES6 packages for buster.

As of ES7, it's bundled with JRE so there's no need to install that one.

Some config has been updated to match what ES is distributed with as of
7.10 (most important of which is config for GC, which has changed in more
recent JRE versions)

Removed a couple of install directives that were about upgrading old
versions - stuff that no longer applies.

And it consumes more memory.

Change-Id: I73f16408b7bb38ed164c909d27978f23faeb3612